pyproject.toml
[tool.poetry]
name = "obsvty"
version = "0.1.0"
description = "Observability and context engineering framework"
authors = ["Fernando Júnior <contato@fernandojr.dev>"]
readme = "README.md"
packages = [{ include = "obsvty", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "2.9.2"
grpcio = "1.76.0"
grpcio-tools = "1.76.0"
grpcio-reflection = "1.76.0"
protobuf = "6.33.0"
pydantic-settings = "^2.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.4.2"
pytest-cov = "5.0.0"
ruff = "0.8.6"
mypy = "1.18.2"
invoke = "2.2.0"
safety = "2.3.5"
pre-commit = "4.3.0"
tomli = "2.0.1"
opentelemetry-proto = "1.38.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"
extend-exclude = [
    "src/obsvty/adapters/messaging/generated",
]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the EOL from the input.
line-ending = "auto"

[tool.mypy]
strict = true
files = ["src"]
exclude = "src/obsvty/adapters/messaging/generated/.*"

[[tool.mypy.overrides]]
module = [
    "google.protobuf",
    "google.protobuf.*",
    "grpc",
    "grpc.*",
    "grpc_health",
    "grpc_health.*",
    "grpc_reflection",
    "grpc_reflection.*",
    "dotenv",
    "dotenv.*",
]
ignore_missing_imports = true
